// src/app/shared/models/ficha-tueste.ts
export interface FichaTueste {
  id_lote: string;
  humedad: number;
  densidad: number;
  caramelizacion: number;
  desarrollo: number;
  temp_desarrollo: number;
  agtrom: number;
  agtrom_gourmet: number;  
  tiempo: number;
  tueste: string;
  id_lote_tostado: string;
  peso_total: number;
  merma: number;  
  merma_gr: number;
}